requirements.txt
phidata
groq
duckduckgo-search
yfinance
python-dotenv